--- src/api/login/index.ts
@@ -1,8 +1,0 @@
-import * as express from 'express'
-import * as loginRouter from './login'
-
-const router = express.Router()
-
-router.use('/', loginRouter)
-
-export = router
