+++ src/api/login/login.schema.ts
@@ -1,0 +1,15 @@
+export const userSchema = {
+  email: {
+    isEmail: true,
+    in: 'body',
+    trim: true,
+    errorMessage: 'Invalid email',
+  },
+  password: {
+    in: 'body',
+    isLength: {
+      errorMessage: 'Password should be at least 5 chars long and max of 10 chars long',
+      options: { min: 5, max: 10 },
+    },
+  },
+}
