+++ src/api/user/delete/delete.ts
@@ -1,0 +1,19 @@
+import * as express from 'express'
+import * as asyncHandler from 'express-async-handler'
+
+import { verifyToken } from '../../../middleware/authentication.middleware'
+import { User } from '../../../services/user.service'
+
+const router = express.Router()
+
+router.delete(
+  '/',
+  verifyToken,
+  asyncHandler(async (req: any, res) => {
+    await User.delete(req.authData.user.user_id)
+
+    res.sendStatus(200)
+  }),
+)
+
+export = router
