+++ src/api/user/index.ts
@@ -1,8 +1,12 @@
 import * as express from 'express'
-import * as userRouter from './user'
+import * as userCreateRouter from './create/create'
+import * as userDeleteRouter from './delete/delete'
+import * as userUpdateRouter from './update/update'
 
 const router = express.Router()
 
-router.use('/', userRouter)
+router.post('/', userCreateRouter)
+router.delete('/', userDeleteRouter)
+router.patch('/', userUpdateRouter)
 
 export = router
